requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
tqdm>=4.65.0
nltk>=3.8.1
requests>=2.31.0
rouge-score>=0.1.2
python-Levenshtein>=0.21.0
